refactor: 입력 금액을 double형이 아닌 Integer형으로 변환하도록 수정

+++ src/main/java/com/precourse/blackjack/domain/util/InputUtil.java
@@ -71,7 +71,7 @@ public class InputUtil {
 		double playerBettingMoney;
 
 		try {
-			playerBettingMoney = Double.parseDouble(reader.readLine());
+			playerBettingMoney = Integer.parseInt(reader.readLine());
 			// checkBettingMoney(playerBettingMoney); 자연수인지 확인
 		} catch (IOException | NumberFormatException e) {
 			throw e;
